requirements.txt
sentence-transformers>=2.7.0
chromadb>=0.5.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
requests>=2.31.0
pyyaml>=6.0
loguru>=0.7.0
mlflow>=2.10.0
numpy>=1.24.0
